Moooooore updates to the code, mainly to use docker compose, and clea…

+++ workshop/expenses-api/utils/expense_operations.py
@@ -1,16 +1,16 @@
+from http.client import HTTPException
 from app.config import database
 from bson.objectid import ObjectId
 from datetime import datetime
 import os
 
-MONGO_DATABASE_NAME = os.getenv("MONGO_DATABASE_NAME", "expensesDb")
+MONGO_DATABASE_NAME = os.getenv("MONGO_DATABASE_NAME", "financeDB")
 
-MONGO_COLLECTION = os.getenv("MONGO_COLLECTION", "expenseDetails")
+MONGO_COLLECTION = os.getenv("MONGO_COLLECTION", "expensesDetails")
 
 class ExpenseModel:
 
     collection = database[MONGO_COLLECTION]
-    deleted_collection = database["ACRONYMS_DELETED_COLLECTION_NAME"]
 
     @classmethod
     async def create(cls, record_data):
@@ -30,13 +30,12 @@ class ExpenseModel:
 
     @classmethod
     async def delete(cls, record_id, reason):
+        if not reason:
+            raise HTTPException(status_code=400, detail="Reason is required for deletion")
+        
         record = await cls.collection.find_one({"expense_id": record_id})
         if not record:
             return None
-
-        # Move to deleted records collection
-        deleted_record = {**record, "deletion_reason": reason}
-        await cls.deleted_collection.insert_one(deleted_record)
 
         await cls.collection.delete_one({"_id": record_id})
         return record
